+++ R/globals.R
@@ -34,7 +34,7 @@ globalVariables(unique(c(
   "tx_sector_descr_de", "voisin", "cp_n_eng", "cp_n_fr", "cp_n_nl", "Gemeentenaam",
   "n_cp_eng", "n_cp_fr", "n_cp_nl", "name", "name_eng", "Nom commune",
   "name_abv", "cp_n_abv", "n_cp_abv", "CP_NAME", ".", "cd_sector_x_31370",
-  "cd_sector_y_31370",
+  "cd_sector_y_31370", "id_regex_belgium_street",
   # phaco_best_data_update : extract_street:
   "key_street_unique",
   # phaco_best_data_update : select_id_street: